Struct libceed::basis::Basis

pub struct Basis { /* fields omitted */ }

Implementations

impl Basis

pub fn create_tensor_H1(
ceed: &Ceed,
dim: usize,
ncomp: usize,
P1d: usize,
Q1d: usize,
interp1d: &[f64],
grad1d: &[f64],
qref1d: &[f64],
qweight1d: &[f64]
) -> Self

pub fn create_tensor_H1_Lagrange(
ceed: &Ceed,
dim: usize,
ncomp: usize,
P: usize,
Q: usize,
qmode: QuadMode
) -> Self

pub fn create_H1(
ceed: &Ceed,
topo: ElemTopology,
ncomp: usize,
nnodes: usize,
nqpts: usize,
interp: &[f64],
grad: &[f64],
qref: &[f64],
qweight: &[f64]
) -> Self

pub fn apply(
&self,
nelem: i32,
tmode: TransposeMode,
emode: EvalMode,
u: &Vector,
v: &mut Vector
)

Apply basis evaluation from nodes to quadrature points or vice versa

nelem - The number of elements to apply the basis evaluation to

tmode - TrasposeMode::NoTranspose to evaluate from nodes to quadrature points, TransposeMode::Transpose to apply the transpose, mapping from quadrature points to nodes

emode - EvalMode::None to use values directly, EvalMode::Interp to use interpolated values, EvalMode::Grad to use gradients, EvalMode::Weight to use quadrature weights

u - Input Vector

v - Output Vector

const q: usize = 6;
let bu = ceed.basis_tensor_H1_Lagrange(1, 1, q, q, QuadMode::GaussLobatto);
let bx = ceed.basis_tensor_H1_Lagrange(1, 1, 2, q, QuadMode::Gauss);

let x_corners = ceed.vector_from_slice(&[-1., 1.]);
let mut x_qpts = ceed.vector(q);
let mut x_nodes = ceed.vector(q);
bx.apply(1, TransposeMode::NoTranspose, EvalMode::Interp,
         &x_corners, &mut x_nodes);
bu.apply(1, TransposeMode::NoTranspose, EvalMode::Interp,
         &x_nodes, &mut x_qpts);

// Create function x^3 + 1 on Gauss Lobatto points
let mut u_arr = [0.; q];
let x_nodes_arr = x_nodes.view();
for i in 0..q {
  u_arr[i] = x_nodes_arr[i]*x_nodes_arr[i]*x_nodes_arr[i] + 1.;
}
let u = ceed.vector_from_slice(&u_arr);

// Map function to Gauss points
let mut v = ceed.vector(q);
v.set_value(0.);
bu.apply(1, TransposeMode::NoTranspose, EvalMode::Interp, &u, &mut v);

// Verify results
let v_arr = v.view();
let x_qpts_arr = x_qpts.view();
for i in 0..q {
  let true_value = x_qpts_arr[i]*x_qpts_arr[i]*x_qpts_arr[i] + 1.;
  assert_eq!(v_arr[i], true_value, "Incorrect basis application");
}

pub fn get_dimension(&self) -> i32

Returns the dimension for given CeedBasis

let dim = 2;
let b = ceed.basis_tensor_H1_Lagrange(dim, 1, 3, 4, QuadMode::Gauss);

let d = b.get_dimension();
assert_eq!(d, dim as i32, "Incorrect dimension");

pub fn get_num_components(&self) -> i32

Returns number of components for given CeedBasis

let ncomp = 2;
let b = ceed.basis_tensor_H1_Lagrange(1, ncomp, 3, 4, QuadMode::Gauss);

let n = b.get_num_components();
assert_eq!(n, ncomp as i32, "Incorrect number of components");

pub fn get_num_nodes(&self) -> i32

Returns total number of nodes (in dim dimensions) of a CeedBasis

let p = 3;
let b = ceed.basis_tensor_H1_Lagrange(2, 1, p, 4, QuadMode::Gauss);

let nnodes = b.get_num_nodes();
assert_eq!(nnodes, (p*p) as i32, "Incorrect number of nodes");

pub fn get_num_quadrature_points(&self) -> i32

Returns total number of quadrature points (in dim dimensions) of a CeedBasis

let q = 4;
let b = ceed.basis_tensor_H1_Lagrange(2, 1, 3, q, QuadMode::Gauss);

let nqpts = b.get_num_quadrature_points();
assert_eq!(nqpts, (q*q) as i32, "Incorrect number of quadrature points");